🐛 fix wrong data type

+++ client/model.go
@@ -40,7 +40,7 @@ type RowAffected struct {
 
 type Score struct {
 	Id    string  `json:"Id"`
-	Score float32 `json:"Score"`
+	Score float64 `json:"Score"`
 }
 
 type User struct {
